requirements.txt
Flask==2.3.2
google-generativeai==0.2.1
transformers==4.34.0
torch==2.0.1
numpy==1.24.3
Pillow==9.5.0
python-dotenv==1.0.0
IPython==8.12.0
